requirements.txt
attrs==23.2.0
black==24.4.2
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
libcst==1.4.0
MarkupSafe==2.1.5
moreorless==0.4.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
PyYAML==6.0.1
stdlibs==2024.5.15
toml==0.10.2
tomlkit==0.12.5
trailrunner==1.4.0
typing_extensions==4.12.2
ufmt==2.6.0
usort==1.0.8.post1
Werkzeug==3.0.3
